application/webapp/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
requests
jinja2
python-multipart
